Make location a session variable to auto populate the search bar after each search
