+++ app/models/business.rb
@@ -39,16 +39,6 @@ class Business < ApplicationRecord
 
   end
 
-  ##only placed here for httparty use
-  def self.get_user_location
-    google_coord_uri = "https://www.googleapis.com/geolocation/v1/geolocate?key=#{ENV['GOOGLE_MAPS_KEY']}"
-    coordinates = HTTParty.post(google_coord_uri, :headers => { 'Content-Type' => 'application/json' }).parsed_response["location"]
-    lat = coordinates["lat"]
-    long = coordinates["lng"]
-    location_city_uri = "https://maps.googleapis.com/maps/api/geocode/json?latlng=#{lat},#{long}&key=#{ENV['GOOGLE_GEO_API']}"
 
-    city = HTTParty.post(location_city_uri, :headers => { 'Content-Type' => 'application/json' }).parsed_response["results"][0]["address_components"][3]["long_name"].to_s
-
-  end
 
 end
