+++ app/models/user.rb
@@ -1,4 +1,9 @@
+require 'httparty'
+require 'uri'
+
 class User < ApplicationRecord
+  include HTTParty
+
   validates :username, presence: true
   # Include default devise modules. Others available are:
   # :confirmable, :lockable, :timeoutable and :omniauthable
@@ -6,4 +11,17 @@ class User < ApplicationRecord
          :recoverable, :rememberable, :trackable, :validatable
   has_many :testimonials
   has_many :votes
+
+
+  def self.get_user_location
+    google_coord_uri = "https://www.googleapis.com/geolocation/v1/geolocate?key=#{ENV['GOOGLE_MAPS_KEY']}"
+    coordinates = HTTParty.post(google_coord_uri, :headers => { 'Content-Type' => 'application/json' }).parsed_response["location"]
+    lat = coordinates["lat"]
+    long = coordinates["lng"]
+    location_city_uri = "https://maps.googleapis.com/maps/api/geocode/json?latlng=#{lat},#{long}&key=#{ENV['GOOGLE_GEO_API']}"
+    city = HTTParty.post(location_city_uri, :headers => { 'Content-Type' => 'application/json' }).parsed_response["results"][0]["address_components"][3]["long_name"].to_s
+  end
+
+
+
 end
